subir archivos modificados

+++ demo/src/main/java/com/example/demo/model/BusinessType.java
@@ -13,7 +13,7 @@ public class BusinessType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_type")
-    private Integer idType;
+    private Long idType;
 
     @Column(name = "name", nullable = false, length = 50, unique = true)
     @NotBlank
@@ -23,8 +23,8 @@ public class BusinessType {
     private String description;
 
     // Getters y setters
-    public Integer getIdType() { return idType; }
-    public void setIdType(Integer idType) { this.idType = idType; }
+    public Long getIdType() { return idType; }
+    public void setIdType(Long idType) { this.idType = idType; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
