+++ demo/src/main/java/com/example/demo/repository/BusinessRepository.java
@@ -1,4 +1,6 @@
 package com.example.demo.repository;
+
 import com.example.demo.model.Business;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface BusinessRepository extends JpaRepository<Business, Integer> { }
+
+public interface BusinessRepository extends JpaRepository<Business, Long> { }
